Separate out fd_txn_m_t
Allows tiles without int128 to use txn_m_t.
